yt-dlp-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-pubsub==2.19.0
python-dotenv==1.0.0
httpx==0.25.2
google-auth==2.23.4
